% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_methods.R
\name{available_methods}
\alias{available_methods}
\title{List dimensionality reduction methods currently available}
\usage{
available_methods()
}
\value{
A character vector of available DR methods
}
\description{
This function shows available methods that worked for \code{dr()} function.
}
\examples{
available_methods()
}
\author{
Lang Zhou
}
