% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element_line2.R
\name{element_line2}
\alias{element_line2}
\title{element_line2}
\usage{
element_line2(
  colour = NULL,
  size = NULL,
  linetype = NULL,
  lineend = NULL,
  color = NULL,
  arrow = NULL,
  inherit.blank = FALSE,
  id,
  xlength = 0.3,
  ylength = 0.3,
  ...
)
}
\arguments{
\item{colour}{line colour}

\item{size}{line size in pts}

\item{linetype}{line type}

\item{lineend}{line end style (round, butt, square)}

\item{color}{aliase to colour}

\item{arrow}{arrow specification, as created by 'grid::arrow()'}

\item{inherit.blank}{whether inherit 'element_blank'}

\item{id}{1 or 2, 1 for axis.line.x.bottom and 2 for axis.line.y.left, only these two axes supported}

\item{xlength}{length of x axis}

\item{ylength}{length of y axis}

\item{...}{additional parameters}
}
\value{
element_line2 object, which is a tailored element_line object
}
\description{
element_line2 for drawing shorten axis lines
}
\author{
Guangchuang Yu
}
