% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_dr}
\alias{theme_dr}
\title{theme_dr}
\usage{
theme_dr(
  xlength = 0.3,
  ylength = 0.3,
  arrow = arrow(length = unit(0.15, "inches"), type = "closed")
)
}
\arguments{
\item{xlength}{length of x axis}

\item{ylength}{length of y axis}

\item{arrow}{arrow specification, as created by 'grid::arrow()'}
}
\description{
Dimensional reduction scatter plot axis theme
}
\author{
Guangchuang Yu
}
