% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_noaxis}
\alias{theme_noaxis}
\title{theme_noaxis}
\usage{
theme_noaxis(...)
}
\arguments{
\item{...}{additional theme setting}
}
\value{
a theme object that disable axes
}
\description{
theme that remove axis
}
\author{
Guangchuang Yu
}
