% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-dr-extract.R
\name{dr_extract}
\alias{dr_extract}
\title{dr_extract}
\usage{
dr_extract(result)
}
\arguments{
\item{result}{DrResult object}
}
\value{
data frame
}
\description{
dr_extract generic
}
\author{
Guangchuang Yu
}
