% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.R
\name{dr}
\alias{dr}
\title{dr}
\usage{
dr(data, fun, ...)
}
\arguments{
\item{data}{input data}

\item{fun}{function to perform dimensional reduction}

\item{...}{additional parameters passed to 'fun'}
}
\value{
DrResult object
}
\description{
dimensional reduction
}
\details{
This function call the user-provided function ('fun') to
perform dimensional reduction on the input data ('data')
}
\author{
Guangchuang Yu
}
