% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.R
\name{dr}
\alias{dr}
\title{dr}
\usage{
dr(data, fun, ...)
}
\arguments{
\item{data}{input data}

\item{fun}{function to perform dimensional reduction}

\item{...}{additional parameters passed to 'fun'}
}
\value{
a DrResult object, which contains 'data' (original data), 'drdata' (coordination after dimensionality reduction), eigenvalue (standard deviation explained by each dimension) and stress (evaluate the effect of dimensionality reduction)
}
\description{
dimensional reduction
}
\details{
This function call the user-provided function ('fun') to
perform dimensional reduction on the input data ('data')
}
\examples{
x = dr(iris[,1:4], prcomp)
autoplot(x, aes(color=.group), metadata=iris$Species)
}
\author{
Guangchuang Yu
}
