% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidydr-package.R
\docType{package}
\name{tidydr-package}
\alias{tidydr}
\alias{tidydr-package}
\title{tidydr: Unify Dimensionality Reduction Results}
\description{
Dimensionality reduction (DR) is widely used in many domain for analyzing and visualizing high-dimensional data. 'tidydr' provides uniform output and is compatible with multiple methods, including 'prcomp', 'mds', 'Rtsne'. etc.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/YuLab-SMU/tidydr/}
  \item Report bugs at \url{https://github.com/YuLab-SMU/tidydr/issues}
}

}
\author{
\strong{Maintainer}: Guangchuang Yu \email{guangchuangyu@gmail.com} (\href{https://orcid.org/0000-0002-6485-8781}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Shuangbin Xu \email{xshuangbin@163.com} (\href{https://orcid.org/0000-0003-3513-5362}{ORCID})
}

Other contributors:
\itemize{
  \item Erqiang Hu \email{13766876214@163.com} [contributor]
}

}
\keyword{internal}
